requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-filter==2.1.0
django-mysql==3.0.0.post1
djangorestframework==3.9.2
html5lib==1.0.1
idna==2.8
Markdown==3.1
mysqlclient==1.4.2.post1
psutil==5.6.1
pytz==2018.9
requests==2.21.0
robobrowser==0.5.3
six==1.12.0
soupsieve==1.9
sqlparse==0.3.0
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.15.2
ojcrawler==1.1.9